requirements.txt
selene>=2.0.0rc4
pytest>=7.2.0
allure-pytest>=2.13.2
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
jsonschema
lxml